cleaned up project
