--- backend/src/main/java/at/htl/db/DELETE_ME.txt
@@ -1,1 +1,0 @@
-Die Ordnerstruktur kann bei Bedarf gelöscht / geändert werden.
