--- backend/src/main/java/at/htl/distributor/DELETE_ME.txt
@@ -1,1 +1,0 @@
-Die Ordnerstruktur kann bei Bedarf gelöscht / geändert werden.
